lib: restructure code to run serializers
The variants of the `op` type (now called `operation`) no longer return
a continuation to call once the operation is complete. Instead it is now
up to the to call the appropriate control operation on the serializer to
update its state before calling `operation` once more. In the case of
writev, the user should now call `shift`. In the case of the yield, the
user can simply call `operation` once the yield is complete.