deque: fix bug in grow logic
The check to determine whether the Deque.t was full was reversed,
causing the underlying array to grow on each enqueue in which the array
was empty, and never when there was not sufficient space. Funny enough,
this growth strategy guarantees that there will always be sufficient
space in the underlying array for the new element to be enqueued, but at
the cost of massive memory usage and, under heavy use, OOM killers. The
check is now correct.