deque: fix memory leak in grow logic
In the case where there is sufficient space in the underlying array fer
new elements, but the free slots are at the front of the array rather
than the back, ensure_space would blit the elements from their current
position to position `0`. This would leave exactly `t.front` pointers at
the end of the array pointing enqueued items, potentially causing a
potential memory leak while the serializer was still active but the
buffer had been written and removed from the front of the queue.